Implemented library support, began working on memory leaks and corruption

+++ src/include/string_tests.c
@@ -3,12 +3,12 @@ int strIsNumeric(char *input) {
     int isNegative = 0;
     if (input[0] == '-') isNegative = 1;
     if (isNegative) {
-        for (int i = 1; i < strlen(input); i++){
+        for (int i = 1; i < (int) strlen(input); i++){
             if (input[i] < '0' || input[i] > '9') return 0;
         }
     }
     else {
-        for (int i = 0; i < strlen(input); i++){
+        for (int i = 0; i < (int) strlen(input); i++){
             if (input[i] < '0' || input[i] > '9') return 0;
         }
     }
